Create generate mod
