+++ Cargo.toml
@@ -15,6 +15,7 @@ openblas = ["blas/openblas", "lapack/openblas"]
 netlib   = ["blas/netlib", "lapack/netlib"]
 
 [dependencies]
+rand = "0.3"
 derive-new = "0.4"
 enum-error-derive = "0.1"
 num-traits  = "0.1"
